Fixes issue #17
The problem was, that later checks returning false for isModified() would overwrite earlier "true"s.
e.g. if gitModified was true, svnModified was null and hgModified was false, the file would be closed
because of the hgModified false passed (hgModified != null && !hgModified) and closed the file although
gitModified was true.

Signed-off-by: Fabian Kreuzkam <fabian@kreuzkam.com>